Oh look, things work again. Who would have thought?
If it's not obvious yet, the problem is that the zlib filter on the incoming
packet causes the contents and size of the reconstructed packet to be
different from the original packet on the wire. So, when we use the
reconstructed packet to seek in the bytestream from the wire, we would end up
in the wrong spot.