chunk: Give damage a threshold.
If damage surpasses the threshold, stop accumulating damage on blocks and
damage the entire chunk instead.

Improves memory usage, CPU usage, and network responsiveness.

The threshold of 176 was determined by looking at the average size of
chunk packets on the wire, which, after compression, are generally much
smaller than equivalent batch block updates. This was just the magic number.
(It's also 2*2*2*2*11, but nobody cares about that.)