inventory: Add a crafting table.
Like a workbench.

This is probably not a good long-term strategy, but it'll help reduce the LOCs
and make things more obvious. Additionally, this really just papers over a
fact I didn't want to state, which is that numpy is *massive* overkill for
this case. So is the dict. Everything, really. :T