plugins: improve performance of winter season transform
To lay snow over a chunk, look at the heighest block in each column, if
it is not snow and not resistant to snow, put a snow block above it.
The change gets rid of the loop over the whole column, and there is no
need to reverse the list anymore, thus improving the speed
significantly.