generators: speed up grass and beaches by using chunk internals
Every call to set_block checks the overall damage of a chunk. This gets
expensive when called a lot of times, like the terrain generators do.
By using chunk internals and setting the block directly, we gain a speedup by
a factor of up to 6.

The OreGenerator doesn't perform better, probably because the cost of
simplex outweighs the cost of using set_block.