entity: Initial support for paintings.
Paintings are loaded from disk and send to clients, they can't be
placed or removed yet.

The create paket, that was send when transfering chunks to clients, was
removed as paintings would not show up otherwise. It seems it is not
needed anyway, pickups continue to work fine.

C: Cherry-picked with merge resolution.

Signed-off-by: Corbin Simpson <MostAwesomeDude@gmail.com>