blocks: Consolidate a few more special blocks.
Strangely enough, if a block is special once, it tends to be special in a few
other ways, too. Saves a few LOCs, and makes it easier to see special blocks
at a glance.

Incidentally, this is starting to look a lot like how Notch stashes his
blocks, although we are (ab)using Python's calling convention to good effect.