chunk: Add a warning on bad indices.
Keeps shitty plugins from trashing the entire generator/seasons/hook pipeline.

And, of course, this was revealed by a bug in Winter.

I wasn't going to do this, but the fact of the matter is that plugins
shouldn't be able to bring the entire chunk generation system to a grinding
halt, and chunks should be more pliant in this regard.