beta/factory: Don't bother scanning chunks if automatons are unloaded.
It's perfectly legal for them to be undefined if the factory's shutting down,
and in that case, we don't want chunks to be scanned since that leads to extra
work. We reach this path if the world's loading chunks, but that might happen
since the world can have chunks requested in order to save last-minute details
about them, like dirty tiles and entities.