world, ibravo: Completely impurify loading from ISerializer.
Objects must now be created in the serializers. This removes a lot of the
pseudo-idempotency from serialization, and permits more flexible serializers
which might store objects instead of serialized data.

Along the way, World stopped being a valid Level, for obvious reasons, so that
had to change as well. That's why this patch is so massive.