Many tests pass!  Also, encryption!
Most of the tests pass again, which is good.  Some tests need to be
rewritten (parsing packets needs to move to protocol testing for
instance) and new content needs tests, too.  The first/hard part of
encryption support has been added -- the encryption packets are now
properly exchanged!  All that remains on that front is encrypting the
streams.