Adding pcap-parser.py to help with debugging.
I wrote a simple script that will parse tcpdump packets to aid in
debugging.  The script assumes that the server and client are being
run on different IP addresses and that no other traffic is in the dump
file.  The framework is flexible enough for a variety of different
parsers, any of which can be selected by modifying the script.
Eventually command-line options will be added.  This script requires
an additional package which is not included in the requirements.txt
files.